// src/shared/api/generated/models/dashboardShellResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardDocumentBodyDto } from "./dashboardDocumentBodyDto";
import type { DashboardIdleWidgetResponseDto } from "./dashboardIdleWidgetResponseDto";

export interface DashboardShellResponseDto {
  dashboardId: string;
  dashboardRevisionId: string;
  document: DashboardDocumentBodyDto;
  revision: number;
  widgets: DashboardIdleWidgetResponseDto[];
}