// src/shared/api/generated/models/reportExportRequestedResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { ReportExportRequestedResponseDtoKind } from "./reportExportRequestedResponseDtoKind";
import type { ReportExportRequestedResponseDtoLane } from "./reportExportRequestedResponseDtoLane";
import type { ReportExportRequestedResponseDtoStatus } from "./reportExportRequestedResponseDtoStatus";

export interface ReportExportRequestedResponseDto {
  exportId: string;
  kind: ReportExportRequestedResponseDtoKind;
  lane: ReportExportRequestedResponseDtoLane;
  savedReportRevisionId: string;
  status: ReportExportRequestedResponseDtoStatus;
}