// src/shared/api/generated/models/reportScheduleChangedResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { ReportScheduleChangedResponseDtoKind } from "./reportScheduleChangedResponseDtoKind";
import type { ReportScheduleChangedResponseDtoStatus } from "./reportScheduleChangedResponseDtoStatus";

export interface ReportScheduleChangedResponseDto {
  kind: ReportScheduleChangedResponseDtoKind;
  scheduleId: string;
  status: ReportScheduleChangedResponseDtoStatus;
  /** @minimum 1 */
  version: number;
}