// src/shared/api/generated/models/reportScheduleResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { ReportScheduleResponseDtoKind } from "./reportScheduleResponseDtoKind";
import type { ReportScheduleResponseDtoStatus } from "./reportScheduleResponseDtoStatus";

export interface ReportScheduleResponseDto {
  kind: ReportScheduleResponseDtoKind;
  scheduleId: string;
  status: ReportScheduleResponseDtoStatus;
  timezone: string;
  /** @minimum 1 */
  version: number;
}