// src/shared/api/generated/models/reportingQueryDefinitionDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { ReportingQueryComparisonDto } from "./reportingQueryComparisonDto";
import type { ReportingQueryFilterDto } from "./reportingQueryFilterDto";
import type { ReportingQueryOrderDto } from "./reportingQueryOrderDto";
import type { ReportingQueryRangeDto } from "./reportingQueryRangeDto";
import type { ReportingQueryDefinitionDtoVersion } from "./reportingQueryDefinitionDtoVersion";

export interface ReportingQueryDefinitionDto {
  comparison?: ReportingQueryComparisonDto;
  datasetRevisionId: string;
  /** @maxItems 16 */
  filters: ReportingQueryFilterDto[];
  /** @maxItems 16 */
  groupBy: string[];
  /**
   * @minimum 1
   * @maximum 10000
   */
  limit?: number;
  /**
   * @minItems 1
   * @maxItems 8
   */
  metrics: string[];
  /** @maxItems 8 */
  order?: ReportingQueryOrderDto[];
  range: ReportingQueryRangeDto;
  version: ReportingQueryDefinitionDtoVersion;
}